move file processor struct to a separate file
